+++ CHANGES.md
@@ -7,6 +7,7 @@ Change Log
 * Added `outlineWidth` support to CZML geometry packets.
 * Added `stroke-width` support to the GeoJSON simple-style implementation.
 * Added `modelMatrix` option to `Primitive` constructor.
+* Fixed a bug in `PolylineGeometry` that would cause the geometry to be split across the IDL for 3D only scenes. [#1197](https://github.com/AnalyticalGraphicsInc/cesium/issues/1197)
 
 ### 1.3 - 2014-11-03